Game Logic correction
Changed the format of the Main and Logic classes so that the correct Piles can be sent to methods for the appropriate logical results.

Method with the same name created about 6 times in Logic class, takes different Objects as inputs thanks to overriding.

Corrected errors and mistakes in code.
Need to delete unused code and implement 3 word commands.