Introduced Drawing Class
This class has methods which allow the cards to be drawn on screen, previously I was using separate methods and drawing the cards based off of a lot of conditions.

I was also using a 1D array, which made navigating around it harder. eg. +-88 would be the position above/below a certain point if printed out in columns of 88.

I have instead switched to a 2D array.

There are now frame methods which draw the cards based off of X,Y co-ordinates. (This is much much more efficient and easier to work with)