Win Condition, Method Visibility, Validity
Changed the win condition, rather than checking if all the board cards are face up and ending game, it instead checks if all the top cards of the foundation piles are "K".

Future implementation for option of "autocomplete" once all the table cards are face up.

Changed method visibility in all classes which needed edits.

All moves are shown to be either valid or invalid.